backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.0
supabase==2.7.4
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-docx==1.1.2
reportlab==4.2.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.0
